simulator/requirements.txt
pycairo
PyGObject
dbus-python
reactivex